STR 2586 update checkpoint status in OL checkpoint db (#1514)
* refactor(db): normalize OL checkpoint storage into payload and signing tables

* feat(db): add observation table and derived checkpoint status

* feat(csm-worker): write checkpoint L1 observation and derive finalization from OL checkpoint DB

* feat(dbtool): derive checkpoint status from normalized OL checkpoint information

+++ crates/chain-worker-new/src/context.rs
@@ -236,7 +236,7 @@ impl ChainWorkerContext for ChainWorkerContextImpl {
         // Get all epoch commitments for this epoch index
         let epoch_commitments = self
             .ol_checkpoint_mgr
-            .get_epoch_commitments_at_blocking(epoch as u64)?;
+            .get_epoch_commitments_at_blocking(epoch)?;
 
         // Fetch the summary for each commitment
         let mut summaries = Vec::with_capacity(epoch_commitments.len());
